decouple the providers file, introduce separate factories
