+++ src/common/database/factory/index.ts
@@ -1,0 +1,2 @@
+export * from './sequelize.factory';
+export * from './migrator.factory';
